requirements.txt
celery==5.1.2
dj-database-url==0.5.0
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
drf-writable-nested==0.6.3
dj-static==0.0.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-decouple==3.4
redis==3.5.3
